Design: 비회원 컴포넌트 추가 (#1)
