+++ src/pages/SignUp.js
@@ -1,4 +1,6 @@
 import React, { useState, useEffect } from "react";
+import { useNavigate } from "react-router-dom";
+import axios from "axios";
 import * as C from "../styles/CommonStyle";
 import * as S from "../styles/SignUpStyle";
 
@@ -33,9 +35,18 @@ function SignUp() {
         });
     };
 
-    const handleSubmit = (e) => {
+    const navigate = useNavigate();
+    const handleSubmit = async (e) => {
         e.preventDefault();
-        // 폼 제출 처리
+        try {
+            const response = await axios.post("api주소/signup", formData);
+            console.log("Response:", response.data);
+            localStorage.setItem("userData", JSON.stringify(formData));
+            navigate("/join");
+        } catch (error) {
+            console.error("회원가입 에러:", error);
+            window.alert("회원가입에 실패했습니다. 다시 시도해주세요.");
+        }
     };
 
     return (
@@ -45,8 +56,8 @@ function SignUp() {
                     <S.Background>
                         <C.PageSpace>
                             <S.SignUp>
+                                <HeaderRegister />
                                 <S.sub_background>
-                                    <HeaderRegister />
                                     <S.title>
                                         <S.logo_title>
                                             <img src={logo} style={{ width: "124px" }}></img>
@@ -56,15 +67,33 @@ function SignUp() {
                                     <S.signup onSubmit={handleSubmit}>
                                         <S.content style={{ marginTop: "73px" }}>
                                             <S.label>ID</S.label>
-                                            <S.input type="text" name="id" placeholder="아이디"></S.input>
+                                            <S.input
+                                                type="text"
+                                                name="id"
+                                                placeholder="아이디"
+                                                value={formData.id}
+                                                onChange={handleChange}
+                                            ></S.input>
                                         </S.content>
                                         <S.content>
                                             <S.label>PW</S.label>
-                                            <S.input type="password" name="pwd" placeholder="비밀번호"></S.input>
+                                            <S.input
+                                                type="password"
+                                                name="pwd"
+                                                placeholder="비밀번호"
+                                                value={formData.pwd}
+                                                onChange={handleChange}
+                                            ></S.input>
                                         </S.content>
                                         <S.content>
                                             <S.label>이름</S.label>
-                                            <S.input type="text" name="name" placeholder="이름"></S.input>
+                                            <S.input
+                                                type="text"
+                                                name="name"
+                                                placeholder="이름"
+                                                value={formData.name}
+                                                onChange={handleChange}
+                                            ></S.input>
                                         </S.content>
                                         <S.content>
                                             <S.label>성별</S.label>
@@ -95,7 +124,7 @@ function SignUp() {
                                         </S.content>
                                         <S.content>
                                             <S.label>연령</S.label>
-                                            <S.select value="" name="age">
+                                            <S.select name="age" value={formData.age} onChange={handleChange}>
                                                 <S.option value="10" name="10">
                                                     10대
                                                 </S.option>
@@ -121,11 +150,17 @@ function SignUp() {
                                         </S.content>
                                         <S.content>
                                             <S.label>휴대전화</S.label>
-                                            <S.input type="tel" name="tel" placeholder="010-0000-0000"></S.input>
+                                            <S.input
+                                                type="tel"
+                                                name="tel"
+                                                placeholder="010-0000-0000"
+                                                value={formData.tel}
+                                                onChange={handleChange}
+                                            ></S.input>
                                         </S.content>
                                         <S.content>
                                             <S.label>가입사유</S.label>
-                                            <S.select value="" name="reason">
+                                            <S.select name="reason" value={formData.reason} onChange={handleChange}>
                                                 <S.option value="interest" name="interest">
                                                     한방에 대한 관심
                                                 </S.option>
